+++ .gitignore
@@ -5,5 +5,4 @@
 /calc_pre.py
 /free.py
 /result
-/result1
-/mnist
+/archive
